Vibrator: Make sure to erase old effect before uploading new effect
The vibration effect need to be erased every time after the playing is
done, and this is implemented in Vibrator::off(). There are cases that
the application may not call Vibrator:off() so the new effect will not
be uploaded successfully. To fix this, make sure the old effect was
erased before uploading a new effect.

Change-Id: Iee4ff44d78646207d3da0a6aa54eefffaf512822